Code achieved at end of 45 minutes
My idea was that we have a hashmap that contains the points and its heights. If they were duplicated, we would take the higher height. If it didn't exist we would see where within the arraylist we should put it, and we would take the heigher height. For the right point we would check to see if it existed, and the take the smaller height.  If it didn't exist we would have taken the height of the point to the the right of it, and take its height.