code created after 35 minutes being reached
The idea of this was to use a priorityQueue to keep track of costs. If duplicated strings were found next to each other, we would add their values to the priorityQueue, and then loop through popping all their values until 1 was left. Then we would iterate from the end of the duplicated substring to avoid repeating.