code achieved in 35 minutes
The problem is straight forward when broken into multiple subproblems. First is to break it up and find the number pairings. Next is to convert to a string, and keep repeating that process until we find n-1. The issue I ran into is that I was stumped on figuring out how to find a datastructure or a clever way of setting up the pairings in a way that number pairings are in order, based on when they were inserted. Once that hurdle is taken care of, then subproblem of converting it into a string is straight forward as well.