Code achieved at end of 35 minutes
This solution gets you really close to the correct answer. The issue it runs into is when the starting string contains duplicate characters, meaning that we skip starting with that string.

Time complexity: O(n^2) since we attempting every combination possible

Space Complexity: O(m) where m is the size of the string