Solution for problem 839
Had some issues starting up the problem, as I didn't know how to begin. After watching a youtube video, approach is similar to a BS in terms of the helper function, and we are testing all combinations of both matrixes to find our max.

Time complexity: O(n^2)
Where n is the size of the array. Time complexity would take longer since we are actually doing 2n^2n, but we can drop the constants.

Space complexity: O(1)
No space was used