More elegant and simplified solution
Problem is able to be solved with just 2 for loops. All it doesn't check the corners since they're always going to be true for them, hence the len-1 checks in the for loops.

Space complexity: O(1) no space was used

Time complexity: O(M x N) where m is the number of rows, and N is the number of columns. For every row we iterate through all the columns to check.