Solution at end of 35 minutes
Error in logic in one of the if statements, because it goes out of bounds. I tried to check for end of boundaries issues, as well as if there was a v and the ball could no longer drop.

If done correctly:
Time complexity: O(M x N) where M is number of rows, and N is number of columns

Space complexity: O(1) no extra space was used