Code achieved at end of 35 minutes
All the checks are in place to see if its valid to check through the entire bank. Issue arises w/ checking if there is a valid mutation w/in the bank.

There is an error in logic within converting the start to end.

If done correctly:

Space complexity: O(n) where n is the size of the bank

Time complexity: O(n) where n is the size of the bank