Code achieved at end of 25 minutes
This can only solve a select few cases. Since solution was done iteratively, we never fully achieve solution for some cases since we don't check the string again to see if its proper. Can be solved properly if done with a while loop to check if there are repeats.

Time complexity: O(n) where n is the length of the string

Space complexity: O(1) we only use basic variables, so no space is being used