Push of code achieved at end of 35 minutes
This solution can solve the problem, but it can run into an issue when the problem size gets too big, because we are constantly, copying over a stack. Because of this, it only passes 88/96 cases.

Space complexity: O(n) where n is the size of the stack

Time complexity: O(n^2) where n is the size of the stack