Correct solution
This solution uses a monotonic stack. If price is less, then we pop from stack and add the values. At end of while loop, we then append that value back to the stack.

Space complexity: O(n)
where n is number of items in the stack

Time complexity: O(n) where n is the number of calls made to the next function